+++ src/test/java/io/appform/dropwizard/sharding/BundleMvccSnapshotTest.java
@@ -1,5 +1,7 @@
 package io.appform.dropwizard.sharding;
 
+import io.appform.dropwizard.sharding.dao.DBShardingBundleBase;
+
 import com.codahale.metrics.health.HealthCheckRegistry;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.appform.dropwizard.sharding.config.ShardedHibernateFactory;
