+++ src/test/java/io/appform/dropwizard/sharding/MultiTenantBundleBasedTestBase.java
@@ -1,5 +1,7 @@
 package io.appform.dropwizard.sharding;
 
+import io.appform.dropwizard.sharding.dao.MultiTenantDBShardingBundleBase;
+
 import com.codahale.metrics.health.HealthCheckRegistry;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.Maps;
