+++ src/test/java/io/appform/dropwizard/sharding/ShardOrderValidationTest.java
@@ -1,5 +1,7 @@
 package io.appform.dropwizard.sharding;
 
+import io.appform.dropwizard.sharding.dao.MultiTenantDBShardingBundleBase;
+
 import io.appform.dropwizard.sharding.config.MultiTenantShardedHibernateFactory;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
