+++ src/test/java/io/appform/dropwizard/sharding/dao/BalancedDBShardingBundleWithNoopBlacklistingStoreTest.java
@@ -15,8 +15,9 @@
  *
  */
 
-package io.appform.dropwizard.sharding;
+package io.appform.dropwizard.sharding.dao;
 
+import io.appform.dropwizard.sharding.*;
 
 import io.appform.dropwizard.sharding.config.ShardedHibernateFactory;
 import io.appform.dropwizard.sharding.dao.testdata.entities.Order;
