+++ src/test/java/io/appform/dropwizard/sharding/dao/DBShardingBundleTestBase.java
@@ -15,7 +15,9 @@
  *
  */
 
-package io.appform.dropwizard.sharding;
+package io.appform.dropwizard.sharding.dao;
+
+import io.appform.dropwizard.sharding.*;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
