+++ src/test/java/io/appform/dropwizard/sharding/dao/EncryptionAtRestTest.java
@@ -1,7 +1,7 @@
 package io.appform.dropwizard.sharding.dao;
 
 import com.google.common.collect.Lists;
-import io.appform.dropwizard.sharding.DBShardingBundleBase;
+import io.appform.dropwizard.sharding.dao.DBShardingBundleBase;
 import io.appform.dropwizard.sharding.ShardInfoProvider;
 import io.appform.dropwizard.sharding.config.ShardingBundleOptions;
 import io.appform.dropwizard.sharding.dao.interceptors.TimerObserver;
