+++ src/test/java/io/appform/dropwizard/sharding/dao/RelationalReadOnlyLockedContextTest.java
@@ -2,7 +2,7 @@ package io.appform.dropwizard.sharding.dao;
 
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
-import io.appform.dropwizard.sharding.DBShardingBundleBase;
+import io.appform.dropwizard.sharding.dao.DBShardingBundleBase;
 import io.appform.dropwizard.sharding.ShardInfoProvider;
 import io.appform.dropwizard.sharding.config.ShardingBundleOptions;
 import io.appform.dropwizard.sharding.dao.interceptors.TimerObserver;
