+++ src/test/java/io/appform/dropwizard/sharding/dao/locktest/LockedContextRollbackTest.java
@@ -2,7 +2,7 @@ package io.appform.dropwizard.sharding.dao.locktest;
 
 import io.appform.dropwizard.sharding.BundleBasedTestBase;
 import io.appform.dropwizard.sharding.DBShardingBundle;
-import io.appform.dropwizard.sharding.DBShardingBundleBase;
+import io.appform.dropwizard.sharding.dao.DBShardingBundleBase;
 import io.appform.dropwizard.sharding.config.ShardedHibernateFactory;
 import io.appform.dropwizard.sharding.dao.LookupDao;
 import io.appform.dropwizard.sharding.dao.RelationalDao;
