+++ src/test/java/io/appform/dropwizard/sharding/observers/BucketKeyPersistorTest.java
@@ -2,7 +2,7 @@ package io.appform.dropwizard.sharding.observers;
 
 import io.appform.dropwizard.sharding.BalancedDBShardingBundle;
 import io.appform.dropwizard.sharding.BundleBasedTestBase;
-import io.appform.dropwizard.sharding.DBShardingBundleBase;
+import io.appform.dropwizard.sharding.dao.DBShardingBundleBase;
 import io.appform.dropwizard.sharding.config.ShardedHibernateFactory;
 import io.appform.dropwizard.sharding.observers.entity.BaseChild;
 import io.appform.dropwizard.sharding.observers.entity.ChildWithDuplicateBucketKey;
