+++ src/test/java/io/appform/dropwizard/sharding/sharding/BucketResolverTest.java
@@ -2,7 +2,7 @@ package io.appform.dropwizard.sharding.sharding;
 
 import io.appform.dropwizard.sharding.BalancedDBShardingBundle;
 import io.appform.dropwizard.sharding.BundleBasedTestBase;
-import io.appform.dropwizard.sharding.DBShardingBundleBase;
+import io.appform.dropwizard.sharding.dao.DBShardingBundleBase;
 import io.appform.dropwizard.sharding.config.ShardedHibernateFactory;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
